server/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
Django==3.1.4
django-channels==0.7.0
idna==2.10
oauthlib==3.1.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
wincertstore==0.2
